@@ -1,15 +1,16 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    
+
     repositories {
         google()
         jcenter()
+        maven { url 'https://jitpack.io' }
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.3.1'
-        
+        classpath 'com.android.tools.build:gradle:3.3.2'
+
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -20,6 +21,7 @@ allprojects {
     repositories {
         google()
         jcenter()
+        maven { url 'https://jitpack.io' }
     }
 }
 
@@ -31,20 +33,20 @@ ext {
     // Sdk and tools
     // Support library and architecture components support minSdk 14 and above.
     minSdkVersion = 14
-    targetSdkVersion = 26
-    compileSdkVersion = 26
-    buildToolsVersion = '26.0.2'
+    targetSdkVersion = 28
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.0'
     // App dependencies
-    supportLibraryVersion = '26.1.0'
+    supportLibraryVersion = '1.0.0-beta01'
     guavaVersion = '18.0'
     junitVersion = '4.12'
     mockitoVersion = '1.10.19'
     powerMockito = '1.6.2'
     hamcrestVersion = '1.3'
-    runnerVersion = '1.0.1'
-    rulesVersion = '1.0.1'
-    espressoVersion = '3.0.1'
-    room_v = '1.1.1'
+    runnerVersion = '1.1.0-alpha4'
+    rulesVersion = '1.1.0-alpha4'
+    espressoVersion = '3.1.0-alpha4'
+    room_v = '2.0.0-beta01'
     retrofit_v = '2.3.0'
-    aac_v = '1.1.1'
+    aac_v = '2.0.0-beta01'
 }
