@@ -1,11 +1,10 @@
 package com.junhyeoklee.som.data.factory;
 
-import android.arch.lifecycle.ViewModel;
-import android.arch.lifecycle.ViewModelProvider;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
 
 import com.junhyeoklee.som.WaterRepository;
-import com.junhyeoklee.som.ui.view_model.AddWaterDateViewModel;
 import com.junhyeoklee.som.ui.view_model.WaterGraphViewModel;
 
 
