@@ -1,19 +1,14 @@
 package com.junhyeoklee.som.data.model;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
+import androidx.room.Update;
 
-import com.junhyeoklee.som.data.model.WaterEntry;
-
-import java.util.Date;
 import java.util.List;
-
-import io.reactivex.Flowable;
 
 @Dao
 public interface WaterDao {
@@ -39,11 +34,11 @@ public interface WaterDao {
     @Update(onConflict = OnConflictStrategy.REPLACE)
     void updateWater(WaterEntry waterEntry);
 
-   @Insert
-   void insertWater(WaterEntry waterEntry);
+    @Insert
+    void insertWater(WaterEntry waterEntry);
 
-   @Delete
-   void deleteWater(WaterEntry waterEntry);
+    @Delete
+    void deleteWater(WaterEntry waterEntry);
 
     @Delete
     void deleteAllWaters(List<WaterEntry> waterEntries);
