@@ -1,9 +1,9 @@
 package com.junhyeoklee.som.ui.activity;
 
 import android.annotation.SuppressLint;
-import android.support.design.internal.BottomNavigationItemView;
-import android.support.design.internal.BottomNavigationMenuView;
-import android.support.design.widget.BottomNavigationView;
+import com.google.android.material.bottomnavigation.BottomNavigationItemView;
+import com.google.android.material.bottomnavigation.BottomNavigationMenuView;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
 import android.util.Log;
 
 import java.lang.reflect.Field;
@@ -21,7 +21,7 @@ class BottomNavigationHelper {
             for (int i = 0; i < menuView.getChildCount(); i++) {
                 BottomNavigationItemView item = (BottomNavigationItemView) menuView.getChildAt(i);
 
-                item.setShiftingMode(false);
+//                item.setShiftingMode(false);
                 // set once again checked value, so view will be updated
 
                 item.setChecked(item.getItemData().isChecked());
