@@ -1,8 +1,8 @@
 package com.junhyeoklee.som.ui.view_model;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 
 import com.junhyeoklee.som.WaterRepository;
 import com.junhyeoklee.som.data.model.WaterEntry;
